+++ changelog.md
@@ -25,6 +25,7 @@
 - Set Kubernetes agent nodes to unscheduled until all agent nodes are not ready.
 - Update golang libraries.
 - Add API to restart agent, datastore and server and update CPU and MEM Ressources [#18](https://github.com/AVENTER-UG/mesos-m3s/pull/18) (thanks to [@itsoksarvesh](https://github.com/itsoksarvesh)).
+- Change datastore healtcheck to Mesos internal check mechanism.
 
 ## v0.4.0